pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "specsplit"
version = "0.1.0"
description = "Disaggregated Speculative Decoding — distributed draft/target LLM pipeline over gRPC"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Samvrit Srinath"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "grpcio-testing>=1.60.0",
    "mypy>=1.8.0",
    "mypy-protobuf>=3.5.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "types-protobuf>=4.25.0",
]

[project.scripts]
specsplit-benchmark = "scripts.benchmark_run:main"

# ---------------------------------------------------------------------------
# Tool Configuration
# ---------------------------------------------------------------------------

[tool.setuptools.packages.find]
include = ["specsplit*"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["specsplit", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["specsplit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring network or model initialization (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.mypy.overrides]
module = "specsplit.proto.*"
ignore_errors = true
